+++ app/components/BookCard.tsx
@@ -4,26 +4,10 @@ import { Image } from "@nextui-org/react";
 import { type Book } from "@/app/data/books";
 import SaveIcon from "@/components/icons/SaveIcon";
 import { useLibraryStore } from "@/store/store";
+import SaveButton from "./SaveButton";
 
 const BookCard = ({ book }: Book) => {
-	const {
-		selectedBooks,
-		addSelectedBook,
-		removeSelectedBook,
-		onOpen,
-		setViewBook,
-	} = useLibraryStore();
-	const saved = selectedBooks.some(
-		(selectedBook) => selectedBook.book.ISBN === book.ISBN
-	);
-
-	const handlePress = () => {
-		if (saved) {
-			removeSelectedBook(book.ISBN);
-		} else {
-			addSelectedBook({ book: book });
-		}
-	};
+	const { onOpen, setViewBook } = useLibraryStore();
 	const handleClick = () => {
 		onOpen();
 		setViewBook({ book: book });
@@ -41,19 +25,7 @@ const BookCard = ({ book }: Book) => {
 						className='overflow-hidden rounded-md object-cover object-center'
 					/>
 					<div className='flex flex-col flex-1 h-full relative'>
-						<Button
-							isIconOnly
-							className='absolute data-[hover]:bg-foreground/10 right-0'
-							radius='full'
-							color='warning'
-							variant='light'
-							onPress={handlePress}
-						>
-							<SaveIcon
-								className={saved ? "[&>path]:stroke-transparent" : ""}
-								fill={saved ? "currentColor" : "none"}
-							/>
-						</Button>
+						<SaveButton book={{ book: book }} />
 						<div className='flex flex-col gap-2 flex-1 '>
 							<h4 className='text-xl font-bold mt-auto'>{book.title}</h4>
 							<p className='text-gray-400 mb-auto'>{book.author?.name}</p>
